Split legacy (RHPAM, BAMOE) and new streams
